varshith/langchain-langraph/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
qdrant-client>=1.12.0
langchain-qdrant>=0.2.0
pypdf>=5.0.0
ollama>=0.4.0
langchain-ollama>=0.2.0
python-dotenv>=1.0.0
streamlit>=1.40.0
pdf2image>=1.17.0
pytesseract>=0.3.10
pillow>=10.4.0
